third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/script-formatter-search.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script src="../../../http/tests/inspector/search/search-test.js"></script>
<script>

function nonFormattedFunction() { var  i = 2 + 2; var a = 4; return a + i; }

var test = function() {
  var scriptSource;
  var shouldRequestContent = false;
  SourcesTestRunner.scriptFormatter().then(startDebuggerTest);
  var scriptFormatter;

  function startDebuggerTest(sf) {
    scriptFormatter = sf;
    SourcesTestRunner.startDebuggerTest(started);
  }

  function started() {
    SourcesTestRunner.showScriptSource('script-formatter-search.html', didShowScriptSource);
  }

  async function didShowScriptSource(frame) {
    scriptSource = frame._uiSourceCode;
    var matches = await scriptSource.searchInContent('magic-string', true, false);
    TestRunner.addResult('Pre-format search results:');
    SourcesTestRunner.dumpSearchMatches(matches);
    shouldRequestContent = true;
    TestRunner.addSniffer(Sources.ScriptFormatterEditorAction.prototype, '_updateButton', uiSourceCodeScriptFormatted);
    scriptFormatter._toggleFormatScriptSource();
  }

  async function uiSourceCodeScriptFormatted() {
    var matches = await scriptSource.searchInContent('magic-string', true, false);
    TestRunner.addResult('Post-format search results:');
    SourcesTestRunner.dumpSearchMatches(matches);
    TestRunner.completeTest();
  }
};

</script>

</head>

<body onload="runTest()">
<p>Tests that search across files works with formatted scripts.
</p>

</body>
</html>
